refactor: sub-package :app by layer (data/service/receiver/sync/widget/ui)
  Move 25 files out of the flat root package into 11 sub-packages
  matching the layered architecture established in earlier phases:

    ui/                MainActivity, Theme, GuardianViewModel
    ui/home, ui/modes, ui/schedules, ui/nfc, ui/info, ui/safety
                       One folder per Compose screen feature
    data/              AppStateRepository, ConfigManager, AppLogger
    service/           BlockerService, ForegroundDetectorService,
                       ForegroundAppDetector, Enforcer, OverlayEnforcer,
                       ForceCloseEnforcer
    receiver/          BootReceiver, ScheduleAlarmReceiver,
                       ServiceRestartReceiver
    sync/              StateSyncer, TimedAlarms
    widget/            GuardianWidget
