+++ app/src/main/java/com/andebugulin/nfcguard/service/ForegroundDetectorService.kt
@@ -1,4 +1,6 @@
-package com.andebugulin.nfcguard
+package com.andebugulin.nfcguard.service
+
+import com.andebugulin.nfcguard.data.AppLogger
 
 import android.accessibilityservice.AccessibilityService
 import android.accessibilityservice.AccessibilityServiceInfo
