+++ app/src/main/java/com/andebugulin/nfcguard/sync/StateSyncer.kt
@@ -1,4 +1,12 @@
-package com.andebugulin.nfcguard
+package com.andebugulin.nfcguard.sync
+
+import com.andebugulin.nfcguard.AppState
+import com.andebugulin.nfcguard.BlockMode
+import com.andebugulin.nfcguard.data.AppLogger
+import com.andebugulin.nfcguard.data.AppStateRepository
+import com.andebugulin.nfcguard.receiver.ScheduleAlarmReceiver
+import com.andebugulin.nfcguard.service.BlockerService
+import com.andebugulin.nfcguard.widget.GuardianWidget
 
 import android.content.Context
