+++ app/src/main/java/com/andebugulin/nfcguard/ui/modes/ModesScreen.kt
@@ -1,4 +1,13 @@
-package com.andebugulin.nfcguard
+package com.andebugulin.nfcguard.ui.modes
+
+import com.andebugulin.nfcguard.ActivationResult
+import com.andebugulin.nfcguard.BlockMode
+import com.andebugulin.nfcguard.Mode
+import com.andebugulin.nfcguard.NfcTag
+import com.andebugulin.nfcguard.Schedule
+import com.andebugulin.nfcguard.ui.GuardianTheme
+import com.andebugulin.nfcguard.ui.GuardianViewModel
+import com.andebugulin.nfcguard.ui.home.HomeScreen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
