+++ app/src/main/java/com/andebugulin/nfcguard/ui/safety/SafeRegimeChallengeDialog.kt
@@ -1,4 +1,6 @@
-package com.andebugulin.nfcguard
+package com.andebugulin.nfcguard.ui.safety
+
+import com.andebugulin.nfcguard.ui.GuardianTheme
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
